Rename SDK packages; remove Python billing
Rename packages/sdk -> packages/sdk-typescript and packages/python-sdk -> packages/sdk-python to align repo layout with language-specific SDKs and update project references and package metadata accordingly. Remove billing SDK surface from the Python SDK: delete billing client and tests, remove billing models/fields and billing exports, and drop the billing namespaces from Relay/AsyncRelay. Update import/path references (AGENTS.md, TS project references, e2e import, tests) so builds and tests remain consistent.
