+++ packages/mcp/tsconfig.json
@@ -8,6 +8,6 @@
   "exclude": ["src/__tests__"],
   "references": [
     { "path": "../types" },
-    { "path": "../sdk" }
+    { "path": "../sdk-typescript" }
   ]
 }
