+++ packages/react/tsconfig.json
@@ -9,6 +9,6 @@
   "exclude": ["src/__tests__"],
   "references": [
     { "path": "../types" },
-    { "path": "../sdk" }
+    { "path": "../sdk-typescript" }
   ]
 }
