+++ packages/sdk-python/src/relay_sdk/__init__.py
@@ -1,7 +1,6 @@
 """Relay SDK — Python client for Relay Transport."""
 
 from .agent import AgentClient, AsyncAgentClient
-from .billing import AsyncBillingClient, BillingClient
 from .client import AsyncHttpClient, HttpClient, SDK_VERSION
 from .errors import RelayError
 from .relay import AsyncRelay, Relay
@@ -12,10 +11,8 @@ __version__ = SDK_VERSION
 __all__ = [
     "AgentClient",
     "AsyncAgentClient",
-    "AsyncBillingClient",
     "AsyncHttpClient",
     "AsyncRelay",
-    "BillingClient",
     "HttpClient",
     "Relay",
     "RelayError",
