+++ packages/sdk-typescript/package.json
@@ -31,7 +31,7 @@
   "repository": {
     "type": "git",
     "url": "https://github.com/AgentWorkforce/relaycast",
-    "directory": "packages/sdk"
+    "directory": "packages/sdk-typescript"
   },
   "files": [
     "dist"
