+++ packages/sdk-typescript/src/casing.ts
