+++ packages/sdk-typescript/src/version.ts
